Membuat hak akses admin dan bumdes

+++ app/Http/Controllers/JenisUsahaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\JenisUsaha;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use DataTables;
 
 class JenisUsahaController extends Controller
@@ -15,20 +16,25 @@ class JenisUsahaController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $data = JenisUsaha::all();
-            return Datatables::of($data)
-                    ->addIndexColumn()
-                    ->addColumn('action', function($row){
-                           $btn = '<a href="javascript:void(0)" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editJenisUsaha"><i class="metismenu-icon pe-7s-pen"></i></a>';
-                           $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteJenisUsaha"><i class="metismenu-icon pe-7s-trash"></i></a>';
-    
-                            return $btn;
-                    })
-                    ->rawColumns(['action'])
-                    ->make(true);
+        if (Auth::user()->hasRole('admin')){
+            if ($request->ajax()) {
+                $data = JenisUsaha::all();
+                return Datatables::of($data)
+                        ->addIndexColumn()
+                        ->addColumn('action', function($row){
+                               $btn = '<a href="javascript:void(0)" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editJenisUsaha"><i class="metismenu-icon pe-7s-pen"></i></a>';
+                               $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteJenisUsaha"><i class="metismenu-icon pe-7s-trash"></i></a>';
+        
+                                return $btn;
+                        })
+                        ->rawColumns(['action'])
+                        ->make(true);
+            }
+            return view('jenisUsaha.index');
         }
-        return view('jenisUsaha.index');
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -38,7 +44,7 @@ class JenisUsahaController extends Controller
      */
     public function create()
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -49,11 +55,16 @@ class JenisUsahaController extends Controller
      */
     public function store(Request $request)
     {
-        $jenis = JenisUsaha::updateOrCreate(
-            ['id' => $request->jenis_id],
-            ['nama' => $request->nama]
-        );
-        return response()->json($jenis);
+        if (Auth::user()->hasRole('admin')){
+            $jenis = JenisUsaha::updateOrCreate(
+                ['id' => $request->jenis_id],
+                ['nama' => $request->nama]
+            );
+            return response()->json($jenis);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -64,7 +75,7 @@ class JenisUsahaController extends Controller
      */
     public function show(JenisUsaha $jenisUsaha)
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -75,8 +86,14 @@ class JenisUsahaController extends Controller
      */
     public function edit($id)
     {
-        $jenis = JenisUsaha::find($id);
-        return response()->json($jenis);
+        if (Auth::user()->hasRole('admin')){
+            $jenis = JenisUsaha::find($id);
+            return response()->json($jenis);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
+        
     }
 
     /**
@@ -88,7 +105,7 @@ class JenisUsahaController extends Controller
      */
     public function update(Request $request, JenisUsaha $jenisUsaha)
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -99,8 +116,13 @@ class JenisUsahaController extends Controller
      */
     public function destroy($id)
     {
-        $jenis = JenisUsaha::find($id);
-        $jenis->delete();
-        return response()->json($jenis);
+        if (Auth::user()->hasRole('admin')){
+            $jenis = JenisUsaha::find($id);
+            $jenis->delete();
+            return response()->json($jenis);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 }
