+++ app/Http/Controllers/KecamatanController.php
@@ -16,7 +16,7 @@ class KecamatanController extends Controller
      */
     public function index()
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -26,8 +26,13 @@ class KecamatanController extends Controller
      */
     public function create($id)
     {
-        $kabupaten = Kabupaten::findOrFail($id);
-        return view('kecamatan.create',compact('kabupaten'));
+        if (Auth::user()->hasRole('admin')){
+            $kabupaten = Kabupaten::findOrFail($id);
+            return view('kecamatan.create',compact('kabupaten'));
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -38,9 +43,14 @@ class KecamatanController extends Controller
      */
     public function store(Request $request)
     {
-        $kecamatan = Kecamatan::create($request->all());
-        $kabupaten = Kabupaten::find($request->get('kabupaten_id'));
-        return redirect()->route('kecamatan.show', $kabupaten)->with('sukses','Kecamatan Berhasil Disimpan');
+        if (Auth::user()->hasRole('admin')){
+            $kecamatan = Kecamatan::create($request->all());
+            $kabupaten = Kabupaten::find($request->get('kabupaten_id'));
+            return redirect()->route('kecamatan.show', $kabupaten)->with('sukses','Kecamatan Berhasil Disimpan');
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -51,25 +61,30 @@ class KecamatanController extends Controller
      */
     public function show(Request $request, $id)
     {
-        $kabupaten = Kabupaten::findOrFail($id);
-        if ($request->ajax()) {
-            $data = Kecamatan::with('bumdes')->where('kabupaten_id', "$id")->get();
-            return Datatables::of($data)
-                    ->addIndexColumn()
-                    ->addColumn('jumlah', function($data) {
-                        return $data->bumdes->count();
-                     })
-                    ->addColumn('action', function($row){
-                           $btn = '<a href="'.route('bumdes.show', $row->id).'" data-toggle="tooltip" title="Lihat Bumdes" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-primary btn-sm editKecamatan"><i class="metismenu-icon pe-7s-info"></i></a>';
-                           $btn = $btn. ' <a href="'.route('kecamatan.edit', $row->id).'" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editKecamatan"><i class="metismenu-icon pe-7s-pen"></i></a>';
-                           $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteKecamatan"><i class="metismenu-icon pe-7s-trash"></i></a>';
-    
-                            return $btn;
-                    })
-                    ->rawColumns(['action'])
-                    ->make(true);
+        if (Auth::user()->hasRole('admin')){
+            $kabupaten = Kabupaten::findOrFail($id);
+            if ($request->ajax()) {
+                $data = Kecamatan::with('bumdes')->where('kabupaten_id', "$id")->get();
+                return Datatables::of($data)
+                        ->addIndexColumn()
+                        ->addColumn('jumlah', function($data) {
+                            return $data->bumdes->count();
+                         })
+                        ->addColumn('action', function($row){
+                               $btn = '<a href="'.route('bumdes.show', $row->id).'" data-toggle="tooltip" title="Lihat Bumdes" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-primary btn-sm editKecamatan"><i class="metismenu-icon pe-7s-info"></i></a>';
+                               $btn = $btn. ' <a href="'.route('kecamatan.edit', $row->id).'" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editKecamatan"><i class="metismenu-icon pe-7s-pen"></i></a>';
+                               $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteKecamatan"><i class="metismenu-icon pe-7s-trash"></i></a>';
+        
+                                return $btn;
+                        })
+                        ->rawColumns(['action'])
+                        ->make(true);
+            }
+            return view('kecamatan.index',compact('kabupaten'));
         }
-        return view('kecamatan.index',compact('kabupaten'));
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -80,8 +95,13 @@ class KecamatanController extends Controller
      */
     public function edit($id)
     {
-        $kecamatan = Kecamatan::findOrFail($id);
-        return view('kecamatan.edit',compact('kecamatan'));  
+        if (Auth::user()->hasRole('admin')){
+            $kecamatan = Kecamatan::findOrFail($id);
+            return view('kecamatan.edit',compact('kecamatan'));  
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -93,10 +113,15 @@ class KecamatanController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $kecamatan = Kecamatan::findOrFail($id);
-        $kecamatan->update($request->all());
-        $kabupaten = Kabupaten::find($request->get('kabupaten_id'));
-        return redirect()->route('kecamatan.show', $kabupaten)->with('sukses','Kecamatan Berhasil Dipdate');
+        if (Auth::user()->hasRole('admin')){
+            $kecamatan = Kecamatan::findOrFail($id);
+            $kecamatan->update($request->all());
+            $kabupaten = Kabupaten::find($request->get('kabupaten_id'));
+            return redirect()->route('kecamatan.show', $kabupaten)->with('sukses','Kecamatan Berhasil Dipdate');
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -107,8 +132,13 @@ class KecamatanController extends Controller
      */
     public function destroy($id)
     {
-        $kecamatan = Kecamatan::find($id);
-        $kecamatan->delete();
-        return response()->json($kecamatan);
+        if (Auth::user()->hasRole('admin')){
+            $kecamatan = Kecamatan::find($id);
+            $kecamatan->delete();
+            return response()->json($kecamatan);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 }
