+++ app/Http/Controllers/UserController.php
@@ -19,20 +19,25 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $data = User::all();
-            return Datatables::of($data)
-                    ->addIndexColumn()
-                    ->addColumn('action', function($row){
-                           $btn = '<a href="javascript:void(0)" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editUser"><i class="metismenu-icon pe-7s-pen"></i></a>';
-                           $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteUser"><i class="metismenu-icon pe-7s-trash"></i></a>';
-    
-                            return $btn;
-                    })
-                    ->rawColumns(['action'])
-                    ->make(true);
+        if (Auth::user()->hasRole('admin')){
+            if ($request->ajax()) {
+                $data = User::all();
+                return Datatables::of($data)
+                        ->addIndexColumn()
+                        ->addColumn('action', function($row){
+                               $btn = '<a href="javascript:void(0)" data-toggle="tooltip" title="Edit" data-id="'.$row->id.'" data-original-title="Edit" class="edit btn btn-info btn-sm editUser"><i class="metismenu-icon pe-7s-pen"></i></a>';
+                               $btn = $btn.' <a href="javascript:void(0)" data-toggle="tooltip" title="Hapus" data-id="'.$row->id.'" data-original-title="Delete" class="btn btn-danger btn-sm deleteUser"><i class="metismenu-icon pe-7s-trash"></i></a>';
+        
+                                return $btn;
+                        })
+                        ->rawColumns(['action'])
+                        ->make(true);
+            }
+            return view('user.index');
         }
-        return view('user.index');
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -42,7 +47,7 @@ class UserController extends Controller
      */
     public function create()
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -53,16 +58,21 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = User::updateOrCreate(
-            ['id' => $request->user_id],
-            [
-                'name' => $request->name,
-                'email' => $request->email,
-                'password' => Hash::make("12345678")
-            ]
-        );
-        $user->assignRole('admin');
-        return response()->json($user);
+        if (Auth::user()->hasRole('admin')){
+            $user = User::updateOrCreate(
+                ['id' => $request->user_id],
+                [
+                    'name' => $request->name,
+                    'email' => $request->email,
+                    'password' => Hash::make("12345678")
+                ]
+            );
+            $user->assignRole('admin');
+            return response()->json($user);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -73,7 +83,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        //
+        return response()->view('errors.404', [abort(404)], 404);
     }
 
     /**
@@ -84,8 +94,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::find($id);
-        return view('user.edit');
+        if (Auth::user()->hasRole('bumdes')){
+            $user = User::find($id);
+            return view('user.edit');
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -97,14 +112,19 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'old_password' => ['required', new MatchOldPassword],
-            'new_password' => ['required'],
-            'confirm_password' => ['same:new_password']
-        ]);
-        User::find(Auth::user()->id)->update(['password' => Hash::make($request->new_password)]);
-        $id_user = ProfilBumdes::find(Auth::user()->profil_bumdes_id);
-        return view('profil', $id_user);
+        if (Auth::user()->hasRole('bumdes')){
+            $request->validate([
+                'old_password' => ['required', new MatchOldPassword],
+                'new_password' => ['required'],
+                'confirm_password' => ['same:new_password']
+            ]);
+            User::find(Auth::user()->id)->update(['password' => Hash::make($request->new_password)]);
+            $id_user = ProfilBumdes::find(Auth::user()->profil_bumdes_id);
+            return view('profil', $id_user);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 
     /**
@@ -115,8 +135,13 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        $user->delete();
-        return response()->json($user);
+        if (Auth::user()->hasRole('admin')){
+            $user = User::find($id);
+            $user->delete();
+            return response()->json($user);
+        }
+        else{
+            return response()->view('errors.403', [abort(403)], 403);
+        }
     }
 }
